streamlit/requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.9
plotly>=5.20.0
pandas>=2.2.0
pdfplumber>=0.11.0
openpyxl>=3.1.2
xlrd>=2.0.1
python-dateutil>=2.9.0
anthropic>=0.32.0
